+++ docs/basic_usage.md
@@ -1,7 +1,7 @@
 # Basic Usage
 
 ### Create an Adjuster
-Create a child class of [`AdjusterBase`][ynabtransactionadjuster.AdjusterBase].
+Create a child class of [`Adjuster`][ynabtransactionadjuster.Adjuster].
 This class needs to implement a `filter()` and an `adjust()` method which contain the intended logic. The `filter()`
 method receives a list of [`OriginalTransaction`][models.OriginalTransaction] objects which can be filtered before 
 adjustement. The `adjust()` method receives a singular [`OriginalTransaction`][models.OriginalTransaction] and a 
@@ -10,11 +10,9 @@ Its attributes can be modified, and it needs to be returned at the end of the fu
 Please check the [detailed usage](detailed_usage.md) section for explanations how to change different attributes.
 
 ```py
-from ynabtransactionadjuster import AdjusterBase
-from ynabtransactionadjuster.models import OriginalTransaction, TransactionModifier
+from ynabtransactionadjuster import Adjuster, OriginalTransaction, TransactionModifier
 
-
-class MyAdjuster(AdjusterBase):
+class MyAdjuster(Adjuster):
 
 	def filter(self, transactions: List[OriginalTransaction]) -> List[OriginalTransaction]:
 		# your implementation
@@ -39,22 +37,18 @@ my_adjuster = MyAdjuster.from_credentials(credentials=my_credentials)
 ```
 
 ### Test
-Test the adjuster on records fetched via the [`dry_run_adjuster()`][functions.dry_run_adjuster] method. The method 
-fetches 
-and executes the adjustments but doesn't write the results back to YNAB. Instead it returns a list of the changed 
-transactions which can be inspected for the changed properties.
+Test the adjuster on records fetched via the `dry_run()` method. It executes the adjustments but doesn't write the 
+results back to YNAB. Instead it returns a list of the changed transactions which can be inspected for the changed 
+properties.
 
 ```py
-from ynabtransactionadjuster import dry_run_adjuster
-
-mod_transactions = dry_run_adjuster(adjuster=my_adjuster, credentials=my_credentials)
+mod_transactions = my_adjuster.dry_run()
 ```
 
 ### Run
-If you are satisfied with the functionality you can execute the adjuster with the [`run_adjuster()`][functions.run_adjuster] 
-method. This will run the adjustments and will update the changed transactions in YNAB. The method 
-returns an integer with the number of successfully updated records.
+If you are satisfied with the functionality you can execute the adjuster with the `run()` method. This will run the 
+adjustments and will update the changed transactions in YNAB. The method returns an integer with the number of 
+successfully updated records.
 ```py
-from ynabtransactionadjuster import run_adjuster
-count_of_updated_transactions = run_adjuster(adjuster=my_adjuster, credentials=my_credentials)
+count_of_updated_transactions = my_adjuster.run()
 ```
